requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI & ML
openai>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0

# Database (Phase 1: Multi-tenancy)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication (Phase 2)
supabase>=2.3.0
python-jose[cryptography]>=3.3.0

# Storage (Phase 1)
boto3>=1.34.0

# Email
sendgrid>=6.11.0

# Testing
pytest>=8.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0.0
click>=8.1.0

# Security & Rate Limiting (Phase 6)
slowapi>=0.1.9

# QR Codes
qrcode[pil]>=7.4.0
